Autogenerate the list of existing validations
Instead of having to manually update the README every time we add a new
validation or update the name of an existing one (potentially causing
merge conflicts on the way), this generates list of existing validations
from their metadata automatically.

So when we add a new validation, it will be added to the docs
automatically based on its name and groups.

Change-Id: Ib5d227d090624ff8844abb5e3fd53c002635705b