Don't rely on overcloudrc
We were reading overcloudrc for the openstack-endpoints validation.
However, the name of tha file may differ if the deployment plan is
called something else than "overcloud" and more importantly, the GUI
deployments don't create the file on the undercloud at all.

This updates the Ansible inventory to read the admin password and
keystone url for the specified plan if available and the validation to
use those values instead of relying on overcloudrc.

Change-Id: I9ff26016c43888267968ef2588f686f7de95bda3
Closes-Bug: #1643808