Drop scapy dependency
This avoids installing a new dependency to run the DHCP validations.
The rogue_dhcp.py script now uses raw sockets to send DHCP Discover
requests and wait for responses.

This commit also changes the timeout to 5 seconds instead of 30 as it
is already plenty of time for a DHCP server to respond.

Change-Id: I2ff2c84601cb63def82b08f3fb670c0f6acd1915