Add lookup plugin for tripleo heat templates
This patch adds a lookup plugin for tripleo heat templates
that playbooks can use as: "{{ lookup('tht') }}"

The lookup will return a list of tuples, each tuple
containing the path and the contents of the file.

To implement this the auth token as well as the plan
name have been added to the inventory.

Change-Id: I8ab1563dc887e51339e669592011f2115cfe5e4c