Show all roles in inventory
This patch updates the tripleo-ansible-inventory to show all roles in
the host list, not just compute and controller nodes.

The retrieval of IPs and stack outputs is slightly refactored: Instead of
retrieving them from Nova, the host IPs are now read from the stacks new
RoleNetIpMap output. The stack outputs are fetched from heat via a new
`StackOutputs` class that ensures outputs are only fetched when needed.

Closes-Bug: #1685544
Closes-Bug: #1689789
Change-Id: Ic8a19f15dd45b383a3f3d403653e3eaef2c1865b