Let undercloud-process-count succeed with warnings
Some services set the number of their processes equal to the number of CPUs,
which poses the risk of running out of memory in cases where this number is
very high. This validation is intended to catch these cases and therefore
breaks if the number of processes is higher than 8.

However, the validation in its current form cannot make a judgement for a
realistic limit for any given situation. This patch changes the validation so
it doesn't fail when the limit is reached but succeeds with warnings.

For a follow-up change we should investigate if increasing the limit might be
reasonable, or even if there might be some heuristics to determine the limit
dynamically based on the undercloud machine's characteristics.

Change-Id: I68e25e2a2f7f840144de7cab97731dde3aeb3694