Add host list by service to inventory
This change adds additional host group sections for each enabled service
to the inventory. Each of these service-based sections contains a
`children` attribute that lists the roles which have the service
enabled. (The stack's `EnabledServices` output is used for this).

The change allows to target validations not just by role but by the
service name, such as "rabbitmq" or "keystone".

Change-Id: I3d1ad63a3d468980cd80685d898e9e7851c3ddb1