Prevent unnecessary calls to dynamic inventory host()
At this time we never return any extra details for hosts from the
host() function, but Ansible still calls it for every host.  Setting
the hostvars key in _meta will prevent this from happening.

Even if we did want to start returning hostvars, we should really
do it with this inventory metadata anyway since it's quite expensive
to retrieve the stack each time the inventory is called.

Change-Id: Ie99422a04e8516c576a38fdb9ffaea439728cc34