Add separate fail-if-no-hosts plugin
Patch #496226 [1] added a sys.exit call to the validation_output plugin
that was fired if a validation was started but no hosts were available
to run it on. This is now separated out into a separate plugin because
it's not really a matter of the output, but rather of the control flow.

[1] https://review.openstack.org/#/c/496226/

Change-Id: Icb364e3dd04f5094782d4a1773bc548f2428afec