Add validation to check VLANs against switch info in Ironic intospection data
This adds a validation to check the VLANs configured in nic config files
against the Ironic introspection data for all nodes.  The introspection process
receives LLDP packets from network switches and stores data including
VLANs, LAG, MTU etc. for all switch ports in the introspection data.
The introspection data in swift is accessed by this validation.

This runs during pre-deployment and, as such, the scheduling of
roles to actual baremetal nodes has not been done yet, so VLANs
cannot be compared against particular switch ports.  This validation
is useful to detect if VLANs have been defined in the template files
but have not been configured on the switch on any switch ports.

A follow-on post-deployment validation can be developed to compare
the VLANs on each role against the switch port VLANs, along with
confirming that bonds are correctly configured on the switch etc.

Change-Id: I5adeefea1534db0ede6993596a0b5d37829f914a