Fix the Network Gateway Validation
The network gateway validation currently breaks if the undercloud.conf
file doesn't exist, due to the missing ``network_gateway`` value.

The validation uses the ignore_missing flag on the undercloud conf
reader, but fails to set its own default for the network gateway value.

This patch adds this default value.

Closes-Bug: #1732468

Change-Id: I47710721323a36616b42455206f779b6f7dda60d