Improve netenv validation and add IPv6 tests
This patch improves the network environment validation to work with very
large IP ranges (when using IPv6).

Prior to this patch the validation would check for every IP within an IP
ranges to see if it was inside (valid) or outside (invalid) of a subnet.
This isn't feasible for many IPv6 cases when IP ranges can be very
large. Instead of comparing each IP this patch checks the
netaddr.IPRange object as a whole against the subnet.

Change-Id: Icbd2e6483c991835d62429d06b5745742186b5c5