Fix the rabbitmq fd limits validation
After containerization of the rabbitmq service the `rabbitmqctl` call that is
made to determine the file descriptor limits doesn't give the expected results
any more (since it's executed outside of the container where the service is
running).

This patch looks up the container id with the `docker ps` command and executes
`rabbitmqctl` inside of the container.

Partial-Bug: #1744322
Change-Id: If9b09c5f43dd434f17135ba632cb6a424fd461db