Fix docker img filter in rabbitmq validation
To determine the current rabbitmq file descriptor limit, a command has to be
executed in the rabbitmq container. The validation uses the `docker ps` command
to find the container ID, but currently uses a name filter that is too strict
to find the right image.

This patch changes the filter to make sure the container image is found.

Change-Id: Ic93411be72314b1ebd10956c003b559ebe14ecde
(cherry picked from commit 38d5d58c9b5c7321fc04293d62b21494dbd85b29)