fix lower constraints
The original patch to set up the lower constraints job used a bad set of
version numbers. It was able to merge because the requirements check job
rules had not yet been updated to require that those settings match the
lower bounds in the requirements lists. This patch fixes the mistakes so
that other patches can change the dependency settings.

Change-Id: Iff383d92f933dc206a02fd78a879a744cbda064c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>