add initial role for containerized undercloud validations
This review contains the first working containerized undercloud
validations, specifically targeted to validate by automation or
ad-hoc by operator, the state of a containerized undercloud.

Initially this checks for the presence of running containers
against a provided list, and the same for the ability to connect
to a given list of TCP ports.

Change-Id: Idafde4a2e7d7ffa82312308dd8b399485373f593