Enhanced checkdisk for undercloud
The main issue with the current check is, we do not take into
account specific locations that can have a dedicated volume.

In order to correct that, we provide a list of locations known
for needing disk space, set a bottom limit, and check what is
actually mounted on the node.
Based on that, we loop on the "mounted volume of interest", and
check the free disk space for each of them.

That way, we ensure we can, at least, deploy the undercloud
without any trouble.

As we have now a fixed list, this can also help operators understanding
WHERE space needs to be allocated, and to what extend.

This patch also take into account the future deprecation of top-level
facts in Ansible.

Please note: the listed "min_size" are the really minimal space needed,
and are probably not sufficient for a productive environment.

Another patch regarding the documentation will be provided once this
change is accepted and merged, in order to reflect this new mode of
calculation.

Change-Id: I71cf319270bfec50304f24a09da34d0ea8b1ecd8