Fix diskspace validation when size is the limit
Previously if you had 25.5G free and the minimum should be 25, the task
would fail. We do not want to fail when the amount of space free is
equal to the minimum required space. We only want to fail when it's less
than the required space. In this case the task needs to be updated to
drop the equals from the failed_when so that if the amount free is equal
to the minimum required amount then it will pass.

Change-Id: I90b4a9617ba00610bee77af84bb1d12511ad452f
Closes-Bug: #1793211