Fix the keystone tokenflush validation
After containerization of the keystone service the `crontab` call that is
made to determine if the 'keystone-manage token_flush' command is present in
the keystone crontab doesn't work anymore (since it's executed outside of the
container where the service is running).

This patch looks up the container id with the `docker ps` command and executes
`crontab` inside of the container.

Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>

Change-Id: I9adaad58a4d53fb3d8f379ffcd3261f1c70e5adb
(cherry picked from commit b1f18c7a2e3856f2b970b0e2fd84e26365584c06)