Fix the undercloud-heat-purge-deleted validation
After containerization of the Heat services the `crontab` call that is
made to determine if the 'heat-manage purge_deleted' command is present in
the Heat crontab doesn't work anymore (since it's executed outside of the
container where the service is running).

This patch looks up the container id with the `docker ps` command and executes
`crontab` inside of the container.

Change-Id: Idbc24b3953c4c3f50a820ba520f90f4a06165281
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>
(cherry picked from commit b69b860563597460184609c61b0e5f358a546ac1)