Fix the undercloud-process-count validation
After containerization of several services the `pgrep` call which should return
the number of running processes doesn't work anymore (since it's executed
outside of the container where the service is running)

This patch executes the `pgrep` command inside of the appropriate container.

Note: The validation in its current form cannot make a judgement for a
realistic limit for any given situation. For that reason, this validation will
always succeed with warning(s).

Change-Id: I0cb33babae2fc285fff7fd6638c578b39e630338
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>