Improve the openshift-nw-openstack validation
This validation should be executed only if an overcloud has been deployed and
working. This patch tests if an overcloud is present and if it's not the case
the validation will succeed with a warning.

Change-Id: I366238859c27d07df4b50a9c0b94400a4731817f
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>
(cherry picked from commit fdf261254eb1dc74c3f2f6a528a175da30d3ee95)