Fix neutron-sanity-check validations
The neutron sanity check validations don't work in a containerized
environment.

This patch runs the neutron-sanity-check command in one of the
neutron containers instead of the host. Also, the number of
configs is updated according to the files available in the
container.

Change-Id: I379d7ec9934cc51d28b676ee2f38cf939d080efe
(cherry picked from commit 1c55753efee67ae6241c0f4a9b3891c944652b63)