Fix ceilometerdb-size validation
Ceilometer doesn't deal with metrics database anymore and Gnocchi doesn't have
interesting parameters to check according to this validation. We can only check
the 'event_ttl_check' in the panko configuration file which crontols the number
of seconds that an event is kept in the database. We also have to note that
panko won't exist anymore in the Stein release.

This patch simplifies this validation.

Change-Id: I2c655cacb748596068d5cfb7cb89eb33ac5cdbbe
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>
(cherry picked from commit 48cdf07abed368205f285493e61f81604f25096d)