Retire the undercloud-selinux-mode validation
Since the containerized undercloud (Rocky Release and above), all the
validations are running from the 'mistral_executor' container. This container
doesn't have and cannot have its own separate SELinux Policies, SELinux will
always appear the be "disabled" in the container, though it is running on the
undercloud host.

This patch retires this validation.

Change-Id: I39cb5e0672c5080ff6d3365e0f56c3e27d19894f
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>