Migrating Undercloud playbooks to their own Ansible Role
The validations should be in the form of Ansible roles, in order to be
easily accessed from the CLI but also from Mistral (as it is currently
the case). It will also allow to get a proper documentation, canvas and
gives the possibility to validate the role before running it by ensuring
there are metadata, output and so on.

Note that all the custom modules/lookup have been copied into the their
top-level directory.

Change-Id: If08678a18165ad12b8fa85f201541fe60c928e1b
Implements: blueprint validation-framework
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>