Improve the network_gateway check on the provisioning network
This patch strengthens the way to get the gateway of the local_subnet in
the undercloud.conf file by getting the local_subnet value first in the
DEFAULT section. And then getting the gateway value from the right
section.

If the local_subnet is not found in the undercloud.conf, then the
default will be 'ctlplane-subnet'.

Change-Id: I618e2fd1ebd1f247264ba835190fb652d0ba0271
Co-Authored-By: Saravanan KR <skramaja@redhat.com>
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>