New validation: ensure services status
This new validation ensures we don't have any failed systemd unit.

Failed units might lead to unexpected errors when we want to update
or upgrade the system, so we have to ensure we don't have anything
in a failed state.

Since this check is pretty generic and useful, it can also be called
as a post-deploy, post-upgrade task. This ensures we have a corret
state AFTER such actions as well.

Change-Id: Id5bb2cce6b4ba41b6c1d860303d9e655ef23f1cf
(cherry picked from commit 1265e2f6675b95bf7487fe041066d8de3164e39e)