Update ceph-health for containers and failure options
Call to 'ceph health' should be wrapped in a container exec.
Display warning for any 'ceph health' status that is not OK.

Add new option, fail_on_ceph_health_err (default: false), to
fail when ceph health is in error.

Add new option, osd_percentage_min (default: 0), to fail
when a certain perecentage of OSDs are not in. When
osd_percentage_min is 0, this check is not performed.

Change-Id: I9a2a60e645443d4c052321c8edc1ec46943b39e9
Implements: blueprint inflight-validations