Ignore ansible-lint issues
We're working on doing a final release of tripleo-validations and the
ansible-lint rules have changed since we last did anything on this
branch. Rather than fixing all the files, let's just ignore the biggest
offending rules so we can do the final release.

204 Lines should be no longer than 160 chars
306 Shells that use pipes should set the pipefail option
602 Don't compare to empty string

I addressed the "206 Variables should have spaces before and after: {{
var_name }}" issue in haproxy.yaml since it was only a single instance.

Change-Id: I96630f3e29f9143fc8dcbbc6aeb8262187023417