Add option to fail if ceph-ansible is not installed
If someone is deploying Ceph then they need ceph-ansible.
The warning might be overlooked and the user may end up
with a message about a missing file. If Ceph is being
deployed then we want to make the validation fail the
deployment if ceph-ansible is not installed so that the
user has a better error message to follow in order to debug.

Also, update warning to refer to the ceph-ansible package.

Change-Id: Ic099c61c74abf6da24a2cce899c944e3bf8ce26b
(cherry picked from commit bf206aa97a5cce5b04de505c294d4d634f05d135)