Improve stackrc file sourcing
We can deploy a TripleO environment from a different user than the
'stack' user. This patch will source the stackrc file coming from the
user home running this script or fail if it doesn't exist.

This patch also enforces the find command to get not more than one
validation found. For example, when running the neutron-sanity-check
playbook, the find command was getting two validations and this script
was running them both.

$ BASEDIR="/usr/share/openstack-tripleo-validations"
$ VALIDATION="neutron-sanity-check"

$ find $BASEDIR/playbooks -type f -regex "^.*${VALIDATION}.y[a]?ml"
/usr/share/openstack-tripleo-validations/playbooks/neutron-sanity-check.yaml
/usr/share/openstack-tripleo-validations/playbooks/undercloud-neutron-sanity-check.yaml

$ find $BASEDIR/playbooks -type f -regex ".*playbooks\/${VALIDATION}.y[a]?ml"
/usr/share/openstack-tripleo-validations/playbooks/neutron-sanity-check.yaml

Change-Id: I889d007fdd757788678fe3d10241024d417632d5
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>