Add pkg names better management between OS releases
The check-latest-minor-version role checks if the latest minor version
of python-tripleoclient rpm is installed on the Undercloud. But now, the
package is called python2-tripleoclient on RHEL/CentOS 7 and
python3-tripleoclient on RHEL 8 and it should be the same on the future
CentOS 8.

This patch fixes the naming issue across the supported OS.

Change-Id: If4aeeca507abc1aee3e783a25fb325e53152753e
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>
(cherry picked from commit 306f531c1c67c9540f36040ff71905dd228d1497)