Add different logic when running against the Undercloud or the Controller
The 'neutron-sanity-check role could be executed against the UC but also
against the Controller nodes. However, the 'container_cli' variable comes
from two different sources of information. This roles should manage this.

OVN is now the default ML2 plugin for the OC nodes while the UC is still
using OVS. Then, this role has to run the 'neutron-sanity-check' command
on two different containers as well:
  - Undercloud -> neutron_ovs_agent
  - Controller -> neutron_api

This patches fixes both of these issues.

Change-Id: I3a027282243db1645361e987cc63f7b6f6fc78ce
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>