Use sorted dictionary before iteration
In Python 3, dict.values() method returns a view of the dictionary's
values but in a different order than Python 2, which returns a
list. This patch sorts the data dictionary first to be sure to get the
same order whatever the Python version.

In Python 3.3 "abstract base classes" in collections (like Mapping or
Sequence) were moved to second-level module collections.abc. So in
Python 3.3+ the real type is collections.abc.Mapping and so
on. Documentation states that the old alias
names (e.g. collections.Mapping) will be available up to Python
3.7 (currently the latest version), however in 3.8 these aliases will be
removed. This patch prepares the code to the future Python 3.8 version.

This patch also fixes the warnings and errors lists to get only unique
messages.

Change-Id: If2c22ad982fafc37ef48b7cd870ea6cf01e7913a
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>