Correct "repos" validation and adds Molecule tests
The way `yum repolist' works isn't consistent on CentOS-7 and Fedora. On
the first, it does fail on a broken repository, spitting an error on
STDERR and exiting with 1.

On Fedora, it sadly just ignores the failing repository with a single
line, and exits with 0.

This patch intends to make the validation more resilient, and adds
Molecule testing in order to ensure it's actually working as expected.

Depends-On: I0551ab56324ede7f1d33e0e9a680c26ea164023c
Change-Id: Ib46f5993eeac6ec39a603da3b8e943c960d7c52d
(cherry picked from commit fa974b9fb9f1e9d3d2ebe5ec5d95095ea018703e)