Correct "repos" validation and adds Molecule tests
The way `yum repolist' works isn't consistent on CentOS-7 and Fedora. On
the first, it does fail on a broken repository, spitting an error on
STDERR and exiting with 1.

On Fedora, it sadly just ignores the failing repository with a single
line, and exits with 0.

This patch intends to make the validation more resilient, and adds
Molecule testing in order to ensure it's actually working as expected.

Change-Id: Ib46f5993eeac6ec39a603da3b8e943c960d7c52d