Add undercloud-key-file and ssh_network options.
To enable the execution via ssh into the Undercloud node it's
required to pass also the ansible_ssh_user private key. This
options is part of the tripleo_common inventory.py, but defaults
to None.
This patch allows to pass this parameter in case it's needed.

Also, this patch allows to pass a different value for the host_network
parameter in TripleoInventory instantiation. Setting it by default to
ctlplane.

Change-Id: Id303f356c1b06906aa5aae43b0032bb80a712508
Related-Bug: #1832932
(cherry picked from commit db30779959c78208b39e716c946152ae8f7e904c)