Switch to promote docs job
We now use a promote job for publishing that avoids building again the
documents after merge. use this here as well.

We can only use the promote jobs when the job run in gate, so use the
same file conditions. To make those easier to sync, get rid of the
special tripleo job and use openstack-tox-docs directly.

For further details about promote jobs, see
http://lists.openstack.org/pipermail/openstack-discuss/2019-August/008610.html

Change-Id: I8432eab9fcec488dc757eabc350ea3e6a9c53102