Fix Python 3 incompatibility in tripleo-ansible-inventory
When running under Python 3, the keys() method on a dict returns a
dict_keys object, not a list. The dict_keys object can't be indexed like
a list, so an exception is thrown. This patch switches from using the
keys() method to using list() on the dict instead, which returns a list
of the keys.

Change-Id: Ifb1814f2406318c4c42edc21581a6fc6d34eb91e