Avoid failure if dhcp_end or dhcp_start isn't set
Since dhcp_start and dhcp_end aren't mandatory option, we have to ensure
the validation doesn't fail when they aren't present.

This change introduce a new "default" param to the "ini" module,
allowing to return a success (with the default value) even if the key
isn't found in the file.

Change-Id: Ibd1e54b2d2748b3bf8d2c1cabad3a9faf832e9b2