Add groups.yaml containing Validation Groups definition
This file will be the source of reference for all the groups used by
tripleo-Validations. It will be deployed next to the playbooks/roles[1]
in order to parsed by the TripleO Validator CLI and its new
subcommand[2].

It will also be the reference for the tripleo-validations documentation
and will be automated in a follow up patch.

[1] - /usr/share/openstack-tripleo-validations/groups.yaml
[2] - openstack tripleo validator group list

Change-Id: I56b9d39c113cfb30695fe8cf2740ae147b0dd3e4
(cherry picked from commit 6df86cdffe3d92b493926140076686c6352c432c)