Add an additional validation to check ceph-ansible repository
In order to address the ceph-ansible potential cross-shipping,
this change adds an extra task to validate the repository that
should be used to install ceph-ansible in the undercloud.
If ceph-ansible is not installed or the repo doesn't match the
specified one, this validation raise an error.

Change-Id: Ib80b55cbd0d4b4dd83d8dbc9dc2afe1d6df2ede5
(cherry picked from commit 329f740d109b87105eac36fade70c69c15781a38)