Add a validation to fail if a Ceph dependency is not installed
When the overcloud is deployed with Ceph we need to install all the
ceph dependencies on the overcloud nodes.
If a ceph dependency is missing, a warning message is displayed and we
can also let the playbook fail setting the related boolean.
Furthermore, if the package is found, the playbook check if a new
version is available.

Closes-Bug: 1855692
Change-Id: I85d09955e839bab35cc06d2f0906f3012c8a1aa9