Ceph PG validation unit shouldn't depend on dict order
The test_simulate_pool_creation_not_enough_osds unit test
was requiring that the error message returned by the function
it tests to contain a dictionary in a certain order. However,
any ordering is valid so the test shouldn't fail if the order
differs. This update of the unit test supports both possible
orderings of the two items.

Change-Id: I80097beb9072c9728de2f08eec4580ea4959be1b