Adds Search Path Override
If you follow the docs for developing validations, you will find
you run into an error when attempting to execute your new validation
because the code is currently hard coded to search in a path in
/usr/share where the packaging installs the validations.  For
development, you must hack the path in the script to make this work.

This patch adds a simple override where a developer can export
a path to the VALIDATIONS_BASEDIR environment variable and that will
be used as the search path for validations.  Example:

cd tripleo-validations
export VALIDATIONS_BASEDIR=$(pwd)

Depends-On: https://review.opendev.org/#/c/706196/
Change-Id: I064c11561af67d284dcf1ab024e0d600e12a8f98