Add basic molecule tests for uncovered roles
This patch adds basic molecule tests which fail until developers create
real ones. The molecule jobs are, of course, non-voting on the check and
not present on the gate. This way, the reminder will be haunting the
damned developers who prefer doing new things, sexier than doing important
molecule tests!

Change-Id: I7ff9506fed0fe4f00e80ff476189f9ea103d81d7
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>