Don't validate token_flush for non-persistent token formats
THT incorporated a change so that the keystone_cron container deploys
only if a persistent token format is used (e.g, `uuid`, `pki`):

  https://review.opendev.org/#/c/682265/

This is because `fernet` and `jws` tokens are completely non-persistent
by way of authenticated encryption and the keystone_cron container only
runs a cron job to execute `keystone-manage token_flush`, which prunes
expired tokens from keystone's token table. This cron job is useful for
deployments relying on `uuid`, `pki`, or some other out-of-tree provider
that requires tokens to be recorded for keystone's token validation
process.

This commit updates the validation to check for the token format and
only performs the validation if the token format isn't `fernet` or
`jws`.

Change-Id: Ib4d0a505021568975b79400d67fd709dd74e9acf