Introducing tripleo_delegate_to on ceph health validation
The purpose of this change is to fix the upgrade scenario
without breaking the existing validation for the
deployment phase.  During the upgrade from queens, the
default 'container_cli' is podman in the undercloud (which
is updated), but it's still docker in the overcloud and this
let the validation playbook to fail because we're computing
this fact starting from inventory_hostname that still points
to the undercloud.  This issue can be solved by removing the
delegation directive from tht and rely on tripleo delegate
wrapper.
This should be enough to access the overcloud (ceph_mon)
variables, setting the right fact.

Closes-Bug: #1871380
Change-Id: I46c4fbe0ef7b281d77bff82483eb0b48b4570bb3
(cherry picked from commit 3baced6e89d4d3af7d2190f4c587a4ee05f5977a)