Add a validation to check the local.
I don't know how to explain how it happens, but we have had several
customers case[1], for example, different local for writing to the database
and reading. What causes an error during minor update, the keystone
client calls the command to list the openstack projects. So we have a
python unicode error. For rabbitmq we force the utf-8 local, so we
decided that we assume that we must be in utf-8 everywhere. Add this
validation to ensure the local system is configure with the good local.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1824513

Change-Id: I8b8f2a17a4963af46dd2c5407fb089414a804247
(cherry picked from commit 6521f1b81adf9502bc2ab10fa72fd75b20348037)
(cherry picked from commit e18073465a5687ff1aca362410ebaa87aa6349cd)