Get UUID from ansible environment variables
Get the uuid set by ansible if exist for logging file creation
on disk.

validations-common is using this uuid set by ansible from ansible-libs
in order to share the same id for logging.
Since the tripleo-validations callback plugin will be used for tripleo
releases for some times, we need to set this here as well

nb: fix doc for pep8

Change-Id: I24ee8fead53dcb9b480447c5fc1d4e2b49e0a8ec