New validation: ensure we have access to the right package version
This validation takes a list of dict describing the packages we want to
verify.
The main goal is to ensure we are avoiding issues at some point with
particularly sensitive packages, such as podman.

We can't use the "package_facts" ansible module, since we're allowing to
check available packages - i.e. versions that aren't installed on the
system.

With the current default, we ensure we get podman 1.6.4 on every nodes
(overcloud and undercloud).
This can be used during an upgrade in order to ensure we're not using
the wrong podman version - either as an inflight validation, or as a
manual step.

Also, it supports the full version number, with the release itself.

This means you can pass "1.6.4-15" for podman, in order to ensure you
get the precise version you want.

Please check the molecule tests in order to see how this validation can
be used.

Please note, the "yum list" part is slow since it will check on the
remote repositories for the available versions!

Last note, we're using "yum" here since the validation might be needed
on pre-dnf releases (namely, centos-7 or rhel-7).

Change-Id: I021a7ad03902ca506885769c1cadc4a449bebbfb
(cherry picked from commit 1c03c9fc56e6861cce3aafbc545abd90cd2dc765)