Allows to use custom role name for validations
The tripleo-ansible-inventory script uses the role names in order to
group the different hosts.

Since we can use custom roles, the fixed "Controller" and others aren't
relevant in such cases.

Ansible allows to use parameter for the "hosts" value in playbooks,
allowing us to call the validations using this command:

openstack tripleo validator run --validation haproxy \
  --extra-vars '{"controller_rolename": "MyCustomController"}'

(or use the --extra-vars-file in order to avoid in-line JSON)

This patch also correct a non-existent group call, Database. The closest
we have is the "mysql" group in the inventory - let's reflect it in
order to avoid useless warning(s).

Change-Id: I3bec039283fe5df56771d84fff5bd5940fd149d8