Refactor vif plugging validations
It makes no sense to lump vif_plugging_timeout and vif_plugging_fatal
into the same tasks. One is a numeric value validated to be larger
than a minimum, the other is a boolean value validated to be set to
True. Split them into separate tasks, and clarify the failure error
messages.

Change-Id: I7a28f6014c55293d018d5fbc3223a26a10550f1c