Add validation to verify the RHSM version.
One of the requirements in the upgrade/update to certain versions
is that the RHSM version configure in the node matches the version
OSP works with. If there is a mismatch on the RHSM version, we would
end up getting wrong packages and finish in an unrecoverable situation.

This validation uses the subscription-manager command to get the
configured release and then checks if it matches the one configured
in check_rhsm_version_expected parameter.

Change-Id: I950395ff8387b01d102fee3a53ad269bb73f16fd