After leapp has been run, verify that the kernel has been upgraded.
This is a new validation to verify the kernel has been upgraded by
checking the kernel version (uname -r) and making sure "el8" appears in it.

Instead of running 'uname -r' - this validation uses the "kernel"
variable from ansible_facts - and checks to see if "el8" is contained
in it.

Change-Id: I3a3c98788a2dfff16412eecaff09f28f7cdef0a6