Docs: Replace tags heading with a real heading
It's not clear to me why the 'Team and repository tags' is heading 1,
but it is clear to me that it takes away from the 'TripleO Validations'
heading.

A quick review of keystone and nova repositories tells me that the
extra heading is not required, so it is replaced with something more
useful.

Change-Id: I4e889d36e9503a75ce8b079f79f20879def402e3