fix mixed up molecule tests
In backporting the molecule tests, the tls-everywhere molecule
tests somehow got mixed up with the switch_vlans molecule test.
This error was masked by the fact that the molecule tests
were not running at all at the time.

This patch sets the content correctly, and re-adds tls-everywhere
to the zuul queues.

Change-Id: I627321bd714bda2c81acf337cea7a7263be6b3ac