[train and before] Define default firewall_driver value
no-op-firewall-nova-driver validation fails if firewall_driver is
not defined in nova.conf. At the same time, default definition is
also OK and validation should pass if firewall_driver is not
defined.

The no-op-firewall-nova-driver validation was removed during Ussuri
cycle[1] so this change is submitted to train and older branches

[1] f1963a4e8396c9b833619e57d4b9509af887c2b0

Closes-Bug: #1901302
Related: rhbz#1891204
Change-Id: I5bf05356bc2bda6ede2df6f7688c2c5e2f46f20e