Make local molecule testing running.
There has been an update in the zuul fact's structure, having now
a projects index which contains the src_dir parameter. This update
was done in run.yaml, however it was missed to be adapted for the
run-local.yml and the pre.yml.

This change adds the fact tripleo_validations_project_path in pre.yml
and updates the zuul dictionary in run-local.yml.

Change-Id: I25ccc7d7a70a64c5abc1a5d621817b2dfad6e38f