Generate inventory without any overcloud
When no overcloud is deployed, we forced to pass the overcloud
hardcoded value to the inventory.
We should, if no --plan or --stack passed, check if there is
plans or stacks on the environment.
If so, used it, if not then just generate undercloud inventory

Depends-On: https://review.opendev.org/c/openstack/tripleo-common/+/766721
Change-Id: Idf4269a2ae5ce9863e1eda2f824e7a9fc93847d2
(cherry picked from commit e9b33aefdc447063b8150342415ae49192875e09)