Introduce tox-ansible to tripleo-validations
This patch adds basic configuration for tox-ansible[1].

Tox-ansible is a plugin for tox which auto-generates tox environments
for running quality assurance tools like ansible-test or molecule. This
tool is rather tightly integrated for the official Molecule testing tool
that integrates with Ansible.

At first, it will allow us to list very quickly all the Ansible Molecule
tests for each roles with their scenario(s):

Example:

    $ tox -l
    ...
    roles-ceph-ceph-ansible-installed
    roles-ceph-default
    roles-check_kernel_version-default
    roles-check_network_gateway-default
    roles-check_rhsm_version-default
    roles-check_rhsm_version-rhsm_mismatch
    roles-check_undercloud_conf-config_OK
    roles-check_undercloud_conf-default
    roles-check_undercloud_conf-deprecated_drivers
    roles-check_undercloud_conf-deprecated_params
    roles-check_undercloud_conf-required_missing
    ...

However, we won't be able to execute them as it for now without to make
some small manual modifications in the molecule.yml file first before to
execute them. This will be covered as soon as possible in a following-patch.

[1] - https://github.com/ansible-community/tox-ansible

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: Ia1606db814f37a11e48508139b90cdd43c083f6b