Add validation to check DockerInsecureRegistryAddress matches the UC hostname
One of the issues identified in the FFU testing was that the
configured DockerInsecureRegistryAddress parameter was not pointing at
the right UC hostname.

Causing issues when retrieving the containers.

This validation retrieve the value for DockerInsecureRegistryAddress
configured in containers-prepare-parameter.yaml file and matches it to
the UC hostname configured in the undercloud.conf.

Co-Authored-By: Daniel Bengtsson <dbengt@redhat.com>
Co-Authored-By: Gaël Chamoulaud <gchamoul@redhat.com>
Change-Id: Iab2e8823996d305d6d8807bb93f4d2f6e1c96906
(cherry picked from commit 5f375a6a0d697a5f5facf18bb7fe1bbbfba35aa5)